requirements.txt
anthropic>=0.40.0
Pillow>=10.0.0
numpy>=1.24.0
pyobjc-framework-Quartz>=10.0
pyobjc-framework-Cocoa>=10.0
rumps>=0.4.0
